# Make refresh_results_cache 6x faster