--- a/evap/evaluation/management/commands/refresh_results_cache.py
+++ b/evap/evaluation/management/commands/refresh_results_cache.py
@@ -3,7 +3,12 @@ from django.core.management.base import BaseCommand
 from django.core.serializers.base import ProgressBar
 
 from evap.evaluation.models import Evaluation
-from evap.results.tools import STATES_WITH_RESULT_TEMPLATE_CACHING, STATES_WITH_RESULTS_CACHING, cache_results
+from evap.results.tools import (
+    GET_RESULTS_PREFETCH_LOOKUPS,
+    STATES_WITH_RESULT_TEMPLATE_CACHING,
+    STATES_WITH_RESULTS_CACHING,
+    cache_results,
+)
 from evap.results.views import warm_up_template_cache
 
 
@@ -19,8 +24,11 @@ class Command(BaseCommand):
         self.stdout.write("Calculating results for all evaluations...")
 
         self.stdout.ending = None
-        evaluations = Evaluation.objects.filter(state__in=STATES_WITH_RESULTS_CACHING)
+        evaluations = Evaluation.objects.filter(state__in=STATES_WITH_RESULTS_CACHING).prefetch_related(
+            *GET_RESULTS_PREFETCH_LOOKUPS,
+        )
         progress_bar = ProgressBar(self.stdout, evaluations.count())
+
         for counter, evaluation in enumerate(evaluations):
             progress_bar.update(counter + 1)
             cache_results(evaluation)