--- a/evap/evaluation/models.py
+++ b/evap/evaluation/models.py
@@ -34,7 +34,7 @@ from evap.evaluation.tools import (
     clean_email,
     date_to_datetime,
     is_external_email,
-    is_prefetched,
+    is_m2m_prefetched,
     translate,
     vote_end_datetime,
 )
@@ -607,7 +607,7 @@ class Evaluation(LoggedModel):
         if self._participant_count is not None:
             return self._participant_count
 
-        if is_prefetched(self, "participants"):
+        if is_m2m_prefetched(self, "participants"):
             return len(self.participants.all())
 
         return self.participants.count()