--- a/evap/evaluation/models_logging.py
+++ b/evap/evaluation/models_logging.py
@@ -3,6 +3,7 @@ import threading
 from collections import defaultdict, namedtuple
 from datetime import date, datetime, time
 from enum import Enum
+from functools import partial
 from json import JSONEncoder
 
 from django.conf import settings
@@ -131,7 +132,7 @@ class LoggedModel(models.Model):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._logentry = None
-        self._m2m_changes = defaultdict(lambda: defaultdict(list))
+        self._m2m_changes = defaultdict(partial(defaultdict, list))  # partial so pickling works
 
     def _as_dict(self):
         """