--- a/evap/results/tools.py
+++ b/evap/results/tools.py
@@ -1,6 +1,7 @@
-from collections import OrderedDict, defaultdict, namedtuple
+from collections import OrderedDict, defaultdict
+from copy import copy
 from math import ceil, modf
-from typing import Dict, List, Tuple, cast
+from typing import Dict, Iterable, List, Optional, Tuple, Union, cast
 
 from django.conf import settings
 from django.core.cache import caches
@@ -16,7 +17,9 @@ from evap.evaluation.models import (
     Questionnaire,
     RatingAnswerCounter,
     TextAnswer,
+    UserProfile,
 )
+from evap.evaluation.tools import discard_cached_related_objects
 
 STATES_WITH_RESULTS_CACHING = {Evaluation.State.EVALUATED, Evaluation.State.REVIEWED, Evaluation.State.PUBLISHED}
 STATES_WITH_RESULT_TEMPLATE_CACHING = {Evaluation.State.PUBLISHED}
@@ -31,45 +34,16 @@ GRADE_COLORS = {
 }
 
 
-class EvaluationResult:
-    def __init__(self, contribution_results):
-        self.contribution_results = contribution_results
-
-    @property
-    def questionnaire_results(self):
-        return [
-            questionnaire_result
-            for contribution_result in self.contribution_results
-            for questionnaire_result in contribution_result.questionnaire_results
-        ]
-
-
-class ContributionResult:
-    def __init__(self, contributor, label, questionnaire_results):
-        self.contributor = contributor
-        self.label = label
-        self.questionnaire_results = questionnaire_results
-
-    @property
-    def has_answers(self):
-        for questionnaire_result in self.questionnaire_results:
-            for question_result in questionnaire_result.question_results:
-                question = question_result.question
-                if question.is_text_question or question.is_rating_question and question_result.has_answers:
-                    return True
-        return False
-
-
-class QuestionnaireResult:
-    def __init__(self, questionnaire, question_results):
-        self.questionnaire = questionnaire
-        self.question_results = question_results
+class TextAnswerVisibility:
+    def __init__(self, visible_by_contribution: Iterable[UserProfile], visible_by_delegation_count: int):
+        self.visible_by_contribution = [discard_cached_related_objects(copy(user)) for user in visible_by_contribution]
+        self.visible_by_delegation_count = visible_by_delegation_count
 
 
 class RatingResult:
     def __init__(self, question, answer_counters, additional_text_result=None):
         assert question.is_rating_question
-        self.question = question
+        self.question = discard_cached_related_objects(copy(question))
         self.additional_text_result = additional_text_result
 
         if answer_counters is not None:
@@ -85,49 +59,98 @@ class RatingResult:
         return CHOICES[self.question.type]
 
     @property
-    def count_sum(self):
+    def count_sum(self) -> Optional[int]:
         if not self.is_published:
             return None
         return sum(self.counts)
 
     @property
-    def minus_balance_count(self):
+    def minus_balance_count(self) -> float:
         assert self.question.is_bipolar_likert_question
         portion_left = sum(self.counts[:3]) + self.counts[3] / 2
         return (self.count_sum - portion_left) / 2
 
     @property
-    def approval_count(self):
+    def approval_count(self) -> Optional[int]:
         assert self.question.is_yes_no_question
         if not self.is_published:
             return None
         return self.counts[0] if self.question.is_positive_yes_no_question else self.counts[1]
 
     @property
-    def average(self):
+    def average(self) -> Optional[float]:
         if not self.has_answers:
             return None
-        return sum(grade * count for count, grade in zip(self.counts, self.choices.grades)) / self.count_sum
+        return sum(grade * count for count, grade in zip(self.counts, self.choices.grades)) / self.count_sum  # type: ignore
 
     @property
-    def has_answers(self):
+    def has_answers(self) -> bool:
         return self.is_published and any(count != 0 for count in self.counts)
 
     @property
-    def is_published(self):
+    def is_published(self) -> bool:
         return self.counts is not None
 
 
 class TextResult:
-    def __init__(self, question, answers, answers_visible_to=None):
+    def __init__(
+        self,
+        question: Question,
+        answers: Iterable[TextAnswer],
+        answers_visible_to: Optional[TextAnswerVisibility] = None,
+    ):
         assert question.can_have_textanswers
-        self.question = question
-        self.answers = answers
+        self.question = discard_cached_related_objects(copy(question))
+        self.answers = [discard_cached_related_objects(copy(answer)) for answer in answers]
         self.answers_visible_to = answers_visible_to
 
 
-HeadingResult = namedtuple("HeadingResult", ("question"))
-TextAnswerVisibility = namedtuple("TextAnswerVisibility", ("visible_by_contribution", "visible_by_delegation_count"))
+class HeadingResult:
+    def __init__(self, question: Question):
+        self.question = discard_cached_related_objects(copy(question))
+
+
+QuestionResult = Union[RatingResult, TextResult, HeadingResult]
+
+
+class QuestionnaireResult:
+    def __init__(self, questionnaire: Questionnaire, question_results: List[QuestionResult]):
+        self.questionnaire = discard_cached_related_objects(copy(questionnaire))
+        self.question_results = question_results
+
+
+class ContributionResult:
+    def __init__(
+        self, contributor: Optional[UserProfile], label: Optional[str], questionnaire_results: List[QuestionnaireResult]
+    ):
+        self.contributor = discard_cached_related_objects(copy(contributor)) if contributor is not None else None
+        self.label = label
+        self.questionnaire_results = questionnaire_results
+
+    @property
+    def has_answers(self) -> bool:
+        for questionnaire_result in self.questionnaire_results:
+            for question_result in questionnaire_result.question_results:
+                question = question_result.question
+                if question.is_text_question:
+                    return True
+                if question.is_rating_question:
+                    assert isinstance(question_result, RatingResult)
+                    return question_result.has_answers
+        return False
+
+
+class EvaluationResult:
+    def __init__(self, contribution_results: List[ContributionResult]):
+        self.contribution_results = contribution_results
+
+    @property
+    def questionnaire_results(self) -> List[QuestionnaireResult]:
+        return [
+            questionnaire_result
+            for contribution_result in self.contribution_results
+            for questionnaire_result in contribution_result.questionnaire_results
+        ]
 
 
 def get_single_result_rating_result(evaluation):
@@ -144,10 +167,10 @@ def get_results_cache_key(evaluation):
     return f"evap.staff.results.tools.get_results-{evaluation.id:d}"
 
 
-def cache_results(evaluation):
+def cache_results(evaluation, *, refetch_related_objects=True):
     assert evaluation.state in STATES_WITH_RESULTS_CACHING
     cache_key = get_results_cache_key(evaluation)
-    caches["results"].set(cache_key, _get_results_impl(evaluation))
+    caches["results"].set(cache_key, _get_results_impl(evaluation, refetch_related_objects=refetch_related_objects))
 
 
 def get_results(evaluation):
@@ -171,27 +194,30 @@ GET_RESULTS_PREFETCH_LOOKUPS = [
 ]
 
 
-def _get_results_impl(evaluation: Evaluation):
+def _get_results_impl(evaluation: Evaluation, *, refetch_related_objects: bool = True):
+    # pylint: disable=too-many-branches, too-many-locals, invalid-name
+    if refetch_related_objects:
+        discard_cached_related_objects(evaluation)
+
     prefetch_related_objects([evaluation], *GET_RESULTS_PREFETCH_LOOKUPS)
 
-    # pylint: disable=too-many-branches
     shown_textanswer_states = [TextAnswer.State.PRIVATE, TextAnswer.State.PUBLISHED]
-    textanswers_per_contribution_question: Dict[(int, int), List[TextAnswer]] = defaultdict(list)
+    textanswers_per_contribution_question: Dict[Tuple[int, int], List[TextAnswer]] = defaultdict(list)
     for contribution in evaluation.contributions.all():
-        for answer in contribution.textanswer_set.all():
-            if answer.state in shown_textanswer_states:
-                textanswers_per_contribution_question[answer.contribution_id, answer.question_id].append(answer)
+        for ta in contribution.textanswer_set.all():
+            if ta.state in shown_textanswer_states:
+                textanswers_per_contribution_question[ta.contribution_id, ta.question_id].append(ta)
 
-    ratinganswercounters_per_contribution_question: Dict[(int, int), List[RatingAnswerCounter]] = defaultdict(list)
+    ratinganswercounters_per_contribution_question: Dict[Tuple[int, int], List[RatingAnswerCounter]] = defaultdict(list)
     for contribution in evaluation.contributions.all():
-        for answer in contribution.ratinganswercounter_set.all():
-            ratinganswercounters_per_contribution_question[answer.contribution_id, answer.question_id].append(answer)
+        for rac in contribution.ratinganswercounter_set.all():
+            ratinganswercounters_per_contribution_question[rac.contribution_id, rac.question_id].append(rac)
 
     contributor_contribution_results = []
     for contribution in evaluation.contributions.all():
         questionnaire_results = []
         for questionnaire in contribution.questionnaires.all():
-            results = []
+            results: List[Union[HeadingResult, TextResult, RatingResult]] = []
             for question in questionnaire.questions.all():
                 if question.is_heading_question:
                     results.append(HeadingResult(question=question))
@@ -209,7 +235,9 @@ def _get_results_impl(evaluation: Evaluation):
                         answer_counters = None
                     results.append(RatingResult(question, answer_counters, additional_text_result=text_result))
                 elif question.is_text_question and evaluation.can_publish_text_results:
+                    assert text_result is not None
                     results.append(text_result)
+
             questionnaire_results.append(QuestionnaireResult(questionnaire, results))
         contributor_contribution_results.append(
             ContributionResult(contribution.contributor, contribution.label, questionnaire_results)