Initially, `./manage.py refresh_results_cache` took approximately 42 seconds for our test data on my machine.

* Prefetching on the `refresh_results_cache` management command brought it down to around 14 seconds (we had around 20'000 db queries) in the process
* Caching the template fragments (the same few fragments were loaded from disk once per render call, totalling in ~9000 file system accesses) brought it down to around 6 seconds (only noticable with DEBUG=True. With debug enabled, this was already cached).

Related to a discussion in #1280:
> you asked me to remind you to keep refresh_results_cache in mind. it currently needs a lot of time on production - so during the update the server is down for several minutes.

6x doesn't sound that impressive, but in comparison, this just flies away:

<img width=50% src="https://user-images.githubusercontent.com/13838962/162849125-4114a84b-0d40-40e8-8c9f-bf87df66faad.gif"/>

<img width=50% src="https://user-images.githubusercontent.com/13838962/162849135-551672f4-0bd1-496e-a091-ada9a7dadfb0.gif"/>

Edit: Current state is even more faster (x8, 4.7s for `refresh_results_cache`)