@He3lixxx 2022-04-13
I don't want this to end up in more code. At the same time, there were a few things that I think we want to handle a bit differently:

* I'd argue it makes sense to put those prefetches into `_get_result_impl`, just before using the values. However, when doing that, `test_finishing_review_updates_results` starts failing -- because with the prefetching, django keeps all the old answers in the "prefetching-cache" (stored in the model's `_prefetched_objects_cache`), even when changing stuff. I think this can be an issue in real code, because we usually call `cache_results` and `get_results` after changing related objects of an evaluation, so the evaluation might contain outdated cached instance. For `create_cache`, our (current) uses seem to be safe in that regard. I tried to go through callers of `get_results` to see if this could actually produce wrong results somewhere, but I wasn't able to come to a conclusion there. I think it would make sense to clear out all django-cached related objects before computing the results.

* Another problem I encountered earlier: With aggressive prefetching, if the objects all contain their cached related objects, these will also be pickled. This means that a TextResult containing a TextAnswer instance will also contain its TextAnswer->Question->Questionnaire->Evaluation->Course->Responsible->Delegates.
This is unnecessary at best and wrong at worst, if these instances are used later after retrieving them from a cache, and they are outdated. We don't have the logic in place to invalidate caches along all hierarchies there. So, I'd argue we don't want to store any related objects in redis.
    * I think this was also a problem before, just not this big, because there were less instances known(?)

* Both combined made me implement `discard_cached_related_objects`, and use it in `_get_results_impl` as well as all the `TextResult`, `RatingResult`, ... classes, on all database instances.
    * Since I'm touching the results anyways, I figured it would make sense to add type annotations. This required reordering them
    * For `HeadingResult` and `TextAnswerVisibility`, adding the conversion-layer means they can no longer be namedtuples. This also requires to adapt one test to use the named accessor (which I'd argue is better anyway)
    * I added `copy()` calls before stripping the related objects, because I think for the average caller, it would be unexpected if the argument was modified.

* For `refresh_results_cache`, we want to prefetch once and use these values (otherwise, it get's ~40% slower). I think the optional kwarg is an okay-ish solution here.
    
* [ ] Tests for `is_m2m_prefetched` (technical debt) and `discard_cached_related_objects` will still be added -- they are necessary, these functions use stuff that's undocumented and not guaranteed to stay available

* [ ] I'll also add a test for https://github.com/e-valuation/EvaP/pull/1738#discussion_r848687453, which hopefully is just a small modification of the existing `TestTextAnswerVisibilityInfo`

Is this too much / are any of these points too cautious?