@He3lixxx 2022-04-18
To make code simpler again: With this, `_get_results_impl` takes
* 30ms for getting the results of a single evaluation, including database queries (before: 180ms)
* 850ms to computing results of all evaluations in test_data.json, with prefetched data
* 4,2s for computing all results of all evaluations, without prefetched data (before: 25s)

so we could think about disabling parts of the results caching, if it makes our lives easier.

For example, for the evaluation_detail results page (based on measurements on a single, random evaluation), it doesn't really matter anymore whether the results are cached or computed (280ms vs 310ms time until the browser gets the response)