@niklasmohrin 2022-04-25
Plus: we talked about introducing some kind of `group_by` utility function for the `defaultdict(list)` dance