@He3lixxx 2022-04-25
Regarding `unordered_groupby`: I ended up not using it in
* results/views.py:395 (`questionnaire_max_answers`): The maximum-computation would be a bit much inside a generator expression
* results/tools.py:373 (`grouped_results`): This uses `extend` instead of `append`, I think making it use append would make it worse.
* results/exporters.py:330 (`questionnaires`): I tried, but the code got longer due to the high amount of nested stuff in the generator expression required
* staff/importers.py:564 (`participants_per_evaluation`): Would get longer. Part of #1634 anyways.