@He3lixxx 2022-04-26
@janno42 I'd love to have the numbers of
```
time ./manage.py refresh_results_cache
```
from before and after this.